requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
Pillow>=8.3.0
numpy>=1.19.0
tqdm>=4.62.0
timm>=0.5.0
transformers>=4.11.0 